+++ src/Output/Output.ts
@@ -1,65 +1,32 @@
 import { Stringable } from '../Stringable/Stringable.ts'
 
-/* eslint @typescript-eslint/no-explicit-any: 0 -- Params can be anything. */
-
 /**
- * Output abstraction for printing messages to output and error streams.
+ * Program output.
  *
  * @since 0.1.0
+ * @since unreleased - Removed all methods except error.
+ * @since unreleased - Added out method.
  */
 export interface Output {
 	/**
-	 * Prints an error to an error stream.
+	 * Outputs a message.
 	 *
-	 * @param  {string | Stringable} message        Message to output.
-	 * @param  {any[]}               optionalParams Optional parameters.
+	 * @param  {string | Stringable}     message Message to output.
+	 * @param  {Record<string, unknown>} context Message context.
 	 * @return {void}
-	 * @since 0.1.0
-	 * @since 0.1.1 - Allow first param to be Stringable.
+	 * @since unreleased
 	 */
-	error(message: string | Stringable, ...optionalParams: any[]): void
+	out(message: string | Stringable, context: Record<string, unknown>): void
 
 	/**
-	 * Prints a warning to an error stream.
+	 * Outputs an error message.
 	 *
-	 * @param  {string | Stringable} message        Message to output.
-	 * @param  {any[]}               optionalParams Optional parameters.
+	 * @param  {string | Stringable}     message Message to output.
+	 * @param  {Record<string, unknown>} context Message context.
 	 * @return {void}
 	 * @since 0.1.0
-	 * @since 0.1.1 - Allow first param to be Stringable.
+	 * @since 0.1.1 - Allow first parameter to be Stringable.
+	 * @since unreleased - Changed second parameter to Record.
 	 */
-	warn(message: string | Stringable, ...optionalParams: any[]): void
-
-	/**
-	 * Prints a log message to an output stream.
-	 *
-	 * @param  {string | Stringable} message        Message to output.
-	 * @param  {any[]}               optionalParams Optional parameters.
-	 * @return {void}
-	 * @since 0.1.0
-	 * @since 0.1.1 - Allow first param to be Stringable.
-	 */
-	log(message: string | Stringable, ...optionalParams: any[]): void
-
-	/**
-	 * Prints an info message to an output stream.
-	 *
-	 * @param  {string | Stringable} message        Message to output.
-	 * @param  {any[]}               optionalParams Optional parameters.
-	 * @return {void}
-	 * @since 0.1.0
-	 * @since 0.1.1 - Allow first param to be Stringable.
-	 */
-	info(message: string | Stringable, ...optionalParams: any[]): void
-
-	/**
-	 * Print a debug message to an output stream.
-	 *
-	 * @param  {string | Stringable} message        Message to output.
-	 * @param  {any[]}               optionalParams Optional parameters.
-	 * @return {void}
-	 * @since 0.1.0
-	 * @since 0.1.1 - Allow first param to be Stringable.
-	 */
-	debug(message: string | Stringable, ...optionalParams: any[]): void
+	error(message: string | Stringable, context: Record<string, unknown>): void
 }